+++ ClassProject/ClassProject/Customer.h
@@ -4,9 +4,9 @@ class Customer
 private:
 
 public:
-	float _moneyOnHand;
-	float _accountValue;
-	Customer(float moneyOnHand, float accountValue);
+	double _moneyOnHand;
+	double _accountValue;
+	Customer(double moneyOnHand, double accountValue);
 	~Customer();
 };
